Tidy up and merge regression tests
