+++ spacy/tests/regression/test_issue3001-3500.py
@@ -88,6 +88,7 @@ def test_issue3199():
     assert list(doc[0:3].noun_chunks) == []
 
 
+@pytest.mark.filterwarnings("ignore::UserWarning")
 def test_issue3209():
     """Test issue that occurred in spaCy nightly where NER labels were being
     mapped to classes incorrectly after loading the model, when the labels
