+++ spacy/tests/regression/test_issue5152.py
@@ -1,6 +1,8 @@
+import pytest
 from spacy.lang.en import English
 
 
+@pytest.mark.filterwarnings("ignore::UserWarning")
 def test_issue5152():
     # Test that the comparison between a Span and a Token, goes well
     # There was a bug when the number of tokens in the span equaled the number of characters in the token (!)
@@ -8,7 +10,6 @@ def test_issue5152():
     text = nlp("Talk about being boring!")
     text_var = nlp("Talk of being boring!")
     y = nlp("Let")
-
     span = text[0:3]  # Talk about being
     span_2 = text[0:3]  # Talk about being
     span_3 = text_var[0:3]  # Talk of being
